BerdCourse(Get_weather)/requirements.txt
requests
pyyaml
pandas
